Refactor package of shardingsphere-distsql-parser-statement (#7999)
* Refactor package of shardingsphere-distsql-parser-statement

* For code format
