+++ shardingsphere-distsql-parser/shardingsphere-distsql-parser-statement/src/main/java/org/apache/shardingsphere/distsql/parser/statement/rdl/CreateShardingRuleStatement.java
@@ -19,6 +19,7 @@ package org.apache.shardingsphere.distsql.parser.statement.rdl;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
+import org.apache.shardingsphere.distsql.parser.segment.rdl.TableRuleSegment;
 
 import java.util.Collection;
